Move host pool class and model type enum out of simulation file to their own header files
